test: move integration tests to a separate file
Move integration tests found in the taskcat-artifact-manager.test.ts
file into a separate file, where replace-in-file is spied on instead of
mocked.

A new "int" folder is created under "__test__" to house tests with spies
and other dependencies. The higher-level folder will be used for true
unit testing, where all dependencies are mocked.

This commit also reverts the mockReplaceInFileSync variable back to a
spy, after it had been changed to a mock. While tests were passing,
backend calls were not actually being made, resulting in false
positives. By keeping the variable as a spy and re-casting it back to
its original type before injecting it into TaskcatArtifactManagerImpl,
we're able to have deeper tests, while maintaining strong types.
